Adds support for non-standard bitrates and timing register adjustments
This update removes the use of hard coded bit timing registers in favor of calculating the registers at runtime. The benefit to this is it allows for any baud rate to be supplied that is within the interfaces ability. It also allows for tweaking of the timing registers to improve network stability and performance. Being able to tweak the timing registers allows for an interface that has a 16MHz clock to work on a network that has devices with 20MHz clocks. Being able to adjust the sample point and the synchronization jump width (SJW) is what makes this possible.

There is only a single api breaking change and that is the baud rate property no longer returns an integer, it returns a `bitrate.Bitrate` instance.

I have removed all of the code that did not pertain specifically to the MCP2515 interface, If at some point this library changes and supports more then one interface chipset family then the code for tht family can be added easily. So if a question gets raised about why something is done the way that it is, keep in mind that it can be added to.